Add files via upload
The phrase "fintech" is gaining popularity and is likely a buzzword that only a select group of tech-savvy individuals or knowledgeable business owners would understand.

Fintech, however, is neither a buzzword nor a concept you can't get familiar with if you stand back and look at it clearly. If you divide it into Finance and Technology, you'll be able to understand it even though it's probably not in your vocabulary yet.

Many people might not be entirely aware of the scope of Fintech app development. The good news is that you shouldn't worry if these topics are unclear to you. Today, we'll talk more about financial technology, developing Fintech applications, and related topics.